* service refactoring
